+++ src/graphql/welcome/schemas.js
@@ -112,7 +112,7 @@ const typeDefs = gql`
     text: String
     textHtml: String
     timestamp: String
-    timestampInt: Int
+    timestampInt: BigInt
     url: String
     project: Project! @relationship(type: "OWNS", direction: IN)
     member: Member! @relationship(type: "DID", direction: IN)
